rename attributes package to annotations
