+++ poc/backend/app/src/main/java/ch/zhaw/studyflow/webserver/sun/SunHttpServerWebServer.java
@@ -5,6 +5,6 @@ import ch.zhaw.studyflow.webserver.WebServer;
 public class SunHttpServerWebServer implements WebServer {
     @Override
     public void start() {
-
+    
     }
 }
